Former-commit-id: 783e31e1f02fea8f9cd5080dc5295eead00ba1ce [formerly 7bf4eb3cc900923a8b3483b45e73333954a9c143] [formerly bd47f3386b2ccaa49ab5a45d874ebcdd653df81f [formerly 60494c4bdb49e4c3cfbdd97f048ee9e5b7b3a81d]]
Former-commit-id: eea148c75240307741b212eff37937ae325d323d [formerly df04c665b6bddd805665558928d8b16938c719fc]
Former-commit-id: f69cd1b0f1762bb2c4a289e553207adbd622ac5a

+++ include/Nazara/Audio/Algorithm.inl
@@ -19,16 +19,16 @@ namespace Nz
 	* \remark The input buffer may be the same as the output one
 	*/
 	template<typename T>
-	void MixToMono(T* input, T* output, unsigned int channelCount, unsigned int frameCount)
+	void MixToMono(T* input, T* output, UInt32 channelCount, UInt64 frameCount)
 	{
 		// To avoid overflow, we use, as an accumulator, a type which is large enough: (u)int 64 bits for integers, double for floatings
 		typedef typename std::conditional<std::is_unsigned<T>::value, UInt64, Int64>::type BiggestInt;
 		typedef typename std::conditional<std::is_integral<T>::value, BiggestInt, double>::type Biggest;
 
-		for (unsigned int i = 0; i < frameCount; ++i)
+		for (UInt64 i = 0; i < frameCount; ++i)
 		{
 			Biggest acc = Biggest(0);
-			for (unsigned int j = 0; j < channelCount; ++j)
+			for (UInt32 j = 0; j < channelCount; ++j)
 				acc += input[i * channelCount + j];
 
 			output[i] = static_cast<T>(acc / channelCount);
